requirements.txt
yangson>=1.6.0
pyyaml>=6.0
jsonpatch>=1.32
requests>=2.25.0  # для тестирования
